[Feature #159965266] Remove snapshot tests and fix action comments
