+++ src/components/auth/LoginForm.jsx
@@ -4,38 +4,26 @@ import { TextInput } from "../../views/Form/index";
 
 const LoginForm = ({ error, password, handleChange, email }) => (
   <form>
-    {error && (
-      <p className="ah-err-msg">
-        {error.email ? "Email is required" : ""}
-        {error ? error.error : ""}
-      </p>
-    )}
-
-    <div className="form-group col-9 m-auto">
+    <div className="form-group col-8 m-auto">
       <TextInput
         type="email"
-        className="form-control mb-3"
+        className="form-control"
         aria-describedby="emailHelp"
         placeholder="Email"
         name="email"
         value={email}
+        error={error.email ? ["Email is required"] : [""]}
         onChange={handleChange}
       />
-
-      {error && (
-        <p className="ah-err-msg">
-          {error.password ? "Password is required" : ""}
-        </p>
-      )}
-
       <div className="form-group">
         <TextInput
           type="password"
-          className="form-control mb-3"
+          className="form-control"
           id="exampleInputPassword1"
           placeholder="Password"
           name="password"
           value={password}
+          error={error.password ? ["Password is required"] : [""]}
           onChange={handleChange}
         />
       </div>
@@ -46,12 +34,8 @@ const LoginForm = ({ error, password, handleChange, email }) => (
 LoginForm.propTypes = {
   email: PropTypes.string.isRequired,
   password: PropTypes.string.isRequired,
-  error: PropTypes.shape(),
+  error: PropTypes.shape({}).isRequired,
   handleChange: PropTypes.func.isRequired
-};
-
-LoginForm.defaultProps = {
-  error: {}
 };
 
 export default LoginForm;
