+++ src/views/App/index.jsx
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import Routes from "../../routes/index";
-import AHHeader from "../Header";
 import "./App.scss";
 
 class App extends Component {
@@ -9,7 +8,6 @@ class App extends Component {
   render() {
     return (
       <div className="App">
-        <AHHeader />
         <Routes />
       </div>
     );
