+++ src/views/Header/index.jsx
@@ -6,7 +6,7 @@ import BaseModal from "../Modal";
 import ForgotPasswordComp from "../../components/auth/ForgotPasswordComp";
 import LoginPage from "../../components/auth/login";
 import AuthButton from "../../components/auth/AuthButton";
-import SignupModal from "../../components/SignupModal";
+import SignupModal from "../../components/auth/SignupModal";
 
 const AHHeader = () => (
   <div>
